Intelligent AI Service Platform (API-First Architecture)/requirements.txt
# Intelligent AI Service Platform - Crafted by CaptainCode

fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
openai==1.3.9
langchain==0.1.0
pinecone-client==3.0.2
celery==5.3.4
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
prometheus-client==0.18.0
python-multipart==0.0.6
PyJWT==2.8.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
aiosqlalchemy==0.4.0
